minor renaming and moving
