+++ src/digitalIdentity-frontend/src/app/components/credDefDetail-table-pop-up/credDefDetail-table-pop-up.component.html
@@ -4,7 +4,7 @@
   </app-filtered-table>
 </div>
 <button mat-raised-button color="primary" (click)="openAddDIWindow()">
-  Add Connection
+  Add DI
 </button>
 <button mat-raised-button color="primary" (click)="close()">
   Close
