+++ src/digitalIdentity-frontend/src/app/components/credDefDetail-table-pop-up/credDefDetail-table-pop-up.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { TablePopUpComponent } from './table-pop-up.component';
+import { TablePopUpComponent } from './credDefDetail-table-pop-up.component';
 
 describe('TablePopUpComponent', () => {
   let component: TablePopUpComponent;
