+++ src/digitalIdentity-frontend/src/app/pages/credential/credential-overview-page/credDef-overview-page.component.ts
@@ -3,7 +3,7 @@ import { Component, isDevMode, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { FilteredTableComponent } from 'src/app/shared/filtered-table/filtered-table.component';
 import { BackendHttpService } from 'src/app/services/backend-http-service/backend-http-service.service';
-import { TablePopUpComponent } from 'src/app/shared/pop-up/table-pop-up/table-pop-up.component';
+import { TablePopUpComponent } from 'src/app/components/credDefDetail-table-pop-up/credDefDetail-table-pop-up.component';
 
 @Component({
   selector: 'app-credential-overview-page',
@@ -46,7 +46,7 @@ export class CredDefOverviewPageComponent implements OnInit {
 
   openAddDIWindow(idx: number,credDefData:any[],dialogRef:MatDialog) {
     if (isDevMode()) {
-      console.log("AddDI")
+      console.log("open AddDI window")
     }
   }
